refactor: 💡 將 ChatRoom 中的 motion 拉出來放，但目前有點擊回應後，並未往下滑到最底的問題存在
