+++ src/components/chat/index.js
@@ -1,5 +1,6 @@
 export * from './BeginMsg';
 export * from './ChatLog';
+export * from './ChatLogMotion';
 export * from './ChatNavBar';
 export * from './SpeakingLoading';
 export * from './ChoiceResponse';
